clash_of_agents/requirements.txt
typer
rich
pydantic
requests
PyYAML
